test(reply): type settled receipt fixtures
