+++ src/auto-reply/reply/dispatch-acp-delivery.test.ts
@@ -6,7 +6,10 @@ import { createAcpDispatchDeliveryCoordinator } from "./dispatch-acp-delivery.js
 import { createReplyDispatcher } from "./reply-dispatcher.js";
 import type { ReplyDispatcher } from "./reply-dispatcher.types.js";
 import { buildTestCtx } from "./test-ctx.js";
-import { createAcpTestConfig } from "./test-fixtures/acp-runtime.js";
+import {
+  createAcpTestConfig,
+  createAcpTestReplyDispatcher as createDispatcher,
+} from "./test-fixtures/acp-runtime.js";
 
 const ttsMocks = vi.hoisted(() => ({
   maybeApplyTtsToPayload: vi.fn(async (paramsUnknown: unknown) => {
@@ -93,38 +96,6 @@ vi.mock("../../channels/plugins/index.js", () => ({
 vi.mock("../../infra/outbound/message-action-runner.js", () => ({
   runMessageAction: (params: unknown) => deliveryMocks.runMessageAction(params),
 }));
-
-function createDispatcher(): ReplyDispatcher {
-  const sendToolResult = vi.fn(() => true);
-  const sendBlockReply = vi.fn(() => true);
-  const sendFinalReply = vi.fn(() => true);
-  const counts = (delivered: number) => ({
-    delivered,
-    deliveredNotVisible: 0,
-    cancelled: 0,
-    failedBeforeSend: 0,
-    failedAfterSend: 0,
-  });
-  return {
-    sendToolResult,
-    sendBlockReply,
-    sendFinalReply,
-    waitForIdle: vi.fn(async () => ({
-      counts: {
-        tool: counts(sendToolResult.mock.calls.length),
-        block: counts(sendBlockReply.mock.calls.length),
-        final: counts(sendFinalReply.mock.calls.length),
-      },
-      anyVisibleDelivered:
-        sendToolResult.mock.calls.length +
-          sendBlockReply.mock.calls.length +
-          sendFinalReply.mock.calls.length >
-        0,
-    })),
-    getQueuedCounts: vi.fn(() => ({ tool: 0, block: 0, final: 0 })),
-    markComplete: vi.fn(),
-  };
-}
 
 function createCoordinator(onReplyStart?: (...args: unknown[]) => Promise<void>) {
   return createAcpDispatchDeliveryCoordinator({
